+++ contrib/install_db4.sh
@@ -96,7 +96,7 @@ echo
 echo "db4 build complete."
 echo
 # shellcheck disable=SC2016
-echo 'When compiling glbitcoind, run `./configure` in the following way:'
+echo 'When compiling bitglobd, run `./configure` in the following way:'
 echo
 echo "  export BDB_PREFIX='${BDB_PREFIX}'"
 # shellcheck disable=SC2016