+++ src/init.cpp
@@ -103,7 +103,7 @@ static const char* FEE_ESTIMATES_FILENAME="fee_estimates.dat";
 /**
  * The PID file facilities.
  */
-static const char* BITCOIN_PID_FILENAME = "glbitcoind.pid";
+static const char* BITCOIN_PID_FILENAME = "bitglobd.pid";
 
 static fs::path GetPidFile()
 {