+++ src/interfaces/node.h
@@ -34,7 +34,7 @@ namespace interfaces {
 class Handler;
 class Wallet;
 
-//! Top-level interface for a bitcoin node (glbitcoind process).
+//! Top-level interface for a bitcoin node (bitglobd process).
 class Node
 {
 public: