+++ test/functional/test_framework/coverage.py
@@ -76,7 +76,7 @@ def get_filename(dirname, n_node):
 
 def write_all_rpc_commands(dirname, node):
     """
-    Write out a list of all RPC functions available in `glbitcoin-cli` for
+    Write out a list of all RPC functions available in `bitglob-cli` for
     coverage comparison. This will only happen once per coverage
     directory.