+++ test/functional/test_framework/test_node.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2017-2019 The Bitcoin Core developers
 # Distributed under the MIT software license, see the accompanying
 # file COPYING or http://www.opensource.org/licenses/mit-license.php.
-"""Class for glbitcoind node under test"""
+"""Class for bitglobd node under test"""
 
 import contextlib
 import decimal
@@ -32,7 +32,7 @@ from .util import (
     p2p_port,
 )
 
-GLBITCOIND_PROC_WAIT_TIMEOUT = 60
+BITGLOBD_PROC_WAIT_TIMEOUT = 60
 
 
 class FailedToStartError(Exception):
@@ -46,7 +46,7 @@ class ErrorMatch(Enum):
 
 
 class TestNode():
-    """A class for representing a glbitcoind node under test.
+    """A class for representing a bitglobd node under test.
 
     This class contains:
 
@@ -59,7 +59,7 @@ class TestNode():
     To make things easier for the test writer, any unrecognised messages will
     be dispatched to the RPC connection."""
 
-    def __init__(self, i, datadir, *, chain, rpchost, timewait, glbitcoind, glbitcoin_cli, coverage_dir, cwd, extra_conf=None, extra_args=None, use_cli=False, start_perf=False):
+    def __init__(self, i, datadir, *, chain, rpchost, timewait, bitglobd, bitglob_cli, coverage_dir, cwd, extra_conf=None, extra_args=None, use_cli=False, start_perf=False):
         """
         Kwargs:
             start_perf (bool): If True, begin profiling the node with `perf` as soon as
@@ -74,7 +74,7 @@ class TestNode():
         self.chain = chain
         self.rpchost = rpchost
         self.rpc_timeout = timewait
-        self.binary = glbitcoind
+        self.binary = bitglobd
         self.coverage_dir = coverage_dir
         self.cwd = cwd
         if extra_conf is not None:
@@ -84,7 +84,7 @@ class TestNode():
         # Note that common args are set in the config file (see initialize_datadir)
         self.extra_args = extra_args
         # Configuration for logging is set as command-line args rather than in the bitcoin.conf file.
-        # This means that starting a glbitcoind using the temp dir to debug a failed test won't
+        # This means that starting a bitglobd using the temp dir to debug a failed test won't
         # spam debug.log.
         self.args = [
             self.binary,
@@ -97,7 +97,7 @@ class TestNode():
             "-uacomment=testnode%d" % i,
         ]
 
-        self.cli = TestNodeCLI(glbitcoin_cli, self.datadir)
+        self.cli = TestNodeCLI(bitglob_cli, self.datadir)
         self.use_cli = use_cli
         self.start_perf = start_perf
 
@@ -163,7 +163,7 @@ class TestNode():
         raise AssertionError(self._node_msg(msg))
 
     def __del__(self):
-        # Ensure that we don't leave any glbitcoind processes lying around after
+        # Ensure that we don't leave any bitglobd processes lying around after
         # the test ends
         if self.process and self.cleanup_on_exit:
             # Should only happen on test failure
@@ -185,7 +185,7 @@ class TestNode():
         if extra_args is None:
             extra_args = self.extra_args
 
-        # Add a new stdout and stderr file each time glbitcoind is started
+        # Add a new stdout and stderr file each time bitglobd is started
         if stderr is None:
             stderr = tempfile.NamedTemporaryFile(dir=self.stderr_dir, delete=False)
         if stdout is None:
@@ -197,7 +197,7 @@ class TestNode():
             cwd = self.cwd
 
         # Delete any existing cookie file -- if such a file exists (eg due to
-        # unclean shutdown), it will get overwritten anyway by glbitcoind, and
+        # unclean shutdown), it will get overwritten anyway by bitglobd, and
         # potentially interfere with our attempt to authenticate
         delete_cookie_file(self.datadir, self.chain)
 
@@ -207,19 +207,19 @@ class TestNode():
         self.process = subprocess.Popen(self.args + extra_args, env=subp_env, stdout=stdout, stderr=stderr, cwd=cwd, **kwargs)
 
         self.running = True
-        self.log.debug("glbitcoind started, waiting for RPC to come up")
+        self.log.debug("bitglobd started, waiting for RPC to come up")
 
         if self.start_perf:
             self._start_perf()
 
     def wait_for_rpc_connection(self):
-        """Sets up an RPC connection to the glbitcoind process. Returns False if unable to connect."""
+        """Sets up an RPC connection to the bitglobd process. Returns False if unable to connect."""
         # Poll at a rate of four times per second
         poll_per_s = 4
         for _ in range(poll_per_s * self.rpc_timeout):
             if self.process.poll() is not None:
                 raise FailedToStartError(self._node_msg(
-                    'glbitcoind exited with status {} during initialization'.format(self.process.returncode)))
+                    'bitglobd exited with status {} during initialization'.format(self.process.returncode)))
             try:
                 rpc = get_rpc_proxy(rpc_url(self.datadir, self.index, self.chain, self.rpchost), self.index, timeout=self.rpc_timeout, coveragedir=self.coverage_dir)
                 rpc.getblockcount()
@@ -239,11 +239,11 @@ class TestNode():
                 # -342 Service unavailable, RPC server started but is shutting down due to error
                 if e.error['code'] != -28 and e.error['code'] != -342:
                     raise  # unknown JSON RPC exception
-            except ValueError as e:  # cookie file not found and no rpcuser or rpcassword. glbitcoind still starting
+            except ValueError as e:  # cookie file not found and no rpcuser or rpcassword. bitglobd still starting
                 if "No RPC credentials" not in str(e):
                     raise
             time.sleep(1.0 / poll_per_s)
-        self._raise_assertion_error("Unable to connect to glbitcoind")
+        self._raise_assertion_error("Unable to connect to bitglobd")
 
     def generate(self, nblocks, maxtries=1000000):
         self.log.debug("TestNode.generate() dispatches `generate` call to `generatetoaddress`")
@@ -303,7 +303,7 @@ class TestNode():
         self.log.debug("Node stopped")
         return True
 
-    def wait_until_stopped(self, timeout=GLBITCOIND_PROC_WAIT_TIMEOUT):
+    def wait_until_stopped(self, timeout=BITGLOBD_PROC_WAIT_TIMEOUT):
         wait_until(self.is_node_stopped, timeout=timeout)
 
     @contextlib.contextmanager
@@ -399,7 +399,7 @@ class TestNode():
 
         if not test_success('readelf -S {} | grep .debug_str'.format(shlex.quote(self.binary))):
             self.log.warning(
-                "perf output won't be very useful without debug symbols compiled into glbitcoind")
+                "perf output won't be very useful without debug symbols compiled into bitglobd")
 
         output_path = tempfile.NamedTemporaryFile(
             dir=self.datadir,
@@ -440,11 +440,11 @@ class TestNode():
     def assert_start_raises_init_error(self, extra_args=None, expected_msg=None, match=ErrorMatch.FULL_TEXT, *args, **kwargs):
         """Attempt to start the node and expect it to raise an error.
 
-        extra_args: extra arguments to pass through to glbitcoind
-        expected_msg: regex that stderr should match when glbitcoind fails
+        extra_args: extra arguments to pass through to bitglobd
+        expected_msg: regex that stderr should match when bitglobd fails
 
-        Will throw if glbitcoind starts without an error.
-        Will throw if an expected_msg is provided and it does not match glbitcoind's stdout."""
+        Will throw if bitglobd starts without an error.
+        Will throw if an expected_msg is provided and it does not match bitglobd's stdout."""
         with tempfile.NamedTemporaryFile(dir=self.stderr_dir, delete=False) as log_stderr, \
              tempfile.NamedTemporaryFile(dir=self.stdout_dir, delete=False) as log_stdout:
             try:
@@ -453,7 +453,7 @@ class TestNode():
                 self.stop_node()
                 self.wait_until_stopped()
             except FailedToStartError as e:
-                self.log.debug('glbitcoind failed to start: %s', e)
+                self.log.debug('bitglobd failed to start: %s', e)
                 self.running = False
                 self.process = None
                 # Check stderr for expected message
@@ -474,9 +474,9 @@ class TestNode():
                                 'Expected message "{}" does not fully match stderr:\n"{}"'.format(expected_msg, stderr))
             else:
                 if expected_msg is None:
-                    assert_msg = "glbitcoind should have exited with an error"
+                    assert_msg = "bitglobd should have exited with an error"
                 else:
-                    assert_msg = "glbitcoind should have exited with expected error " + expected_msg
+                    assert_msg = "bitglobd should have exited with expected error " + expected_msg
                 self._raise_assertion_error(assert_msg)
 
     def add_p2p_connection(self, p2p_conn, *, wait_for_verack=True, **kwargs):
@@ -531,7 +531,7 @@ def arg_to_cli(arg):
         return str(arg)
 
 class TestNodeCLI():
-    """Interface to glbitcoin-cli for an individual node"""
+    """Interface to bitglob-cli for an individual node"""
 
     def __init__(self, binary, datadir):
         self.options = []
@@ -541,7 +541,7 @@ class TestNodeCLI():
         self.log = logging.getLogger('TestFramework.bitcoincli')
 
     def __call__(self, *options, input=None):
-        # TestNodeCLI is callable with glbitcoin-cli command-line options
+        # TestNodeCLI is callable with bitglob-cli command-line options
         cli = TestNodeCLI(self.binary, self.datadir)
         cli.options = [str(o) for o in options]
         cli.input = input
@@ -560,17 +560,17 @@ class TestNodeCLI():
         return results
 
     def send_cli(self, command=None, *args, **kwargs):
-        """Run glbitcoin-cli command. Deserializes returned string as python object."""
+        """Run bitglob-cli command. Deserializes returned string as python object."""
         pos_args = [arg_to_cli(arg) for arg in args]
         named_args = [str(key) + "=" + arg_to_cli(value) for (key, value) in kwargs.items()]
-        assert not (pos_args and named_args), "Cannot use positional arguments and named arguments in the same glbitcoin-cli call"
+        assert not (pos_args and named_args), "Cannot use positional arguments and named arguments in the same bitglob-cli call"
         p_args = [self.binary, "-datadir=" + self.datadir] + self.options
         if named_args:
             p_args += ["-named"]
         if command is not None:
             p_args += [command]
         p_args += pos_args + named_args
-        self.log.debug("Running glbitcoin-cli command: %s" % command)
+        self.log.debug("Running bitglob-cli command: %s" % command)
         process = subprocess.Popen(p_args, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
         cli_stdout, cli_stderr = process.communicate(input=self.input)
         returncode = process.poll()